Added client side validation of add/edit section
No specific server side validation is done at this point, but I think
pygit2 might object to some of the error conditions.  This is something
for later.

Currently the client side validation just pops up an alert in the middle
of the screen above the content boxes.  This will do for now, but in the
future it'd be good to highlight the offending fields or show tooltips
on them or something.  That's also something for later