Main view now updates when history item selected
Previously a fetch was occuring when the section history item was
selected, but the server side was not yet implemented.

This changeset updates the server side to support fetching of individual
sections, and some tests for them.

Also updates the sectionItemView to call render when the model is
changed (quite surprised this wasn't a marionette default, but suppose
it's not what you'd want all the time)