Added login view & template
This changeset adds a view for the login page, a corresponding
template and a simple unit test.
Added images & css for use with the "Login using ..." buttons

Also temporarily added some redirect code to app.coffee that will
cause the index page to redirect to login every time.  This code can be
replaced with something more suitable after development of login has
progressed a bit.