Navbar 'active' now reflects current page
Previously the navBar would always act as if "Overall Layout" was the
current page.  This changeset updates it to reflect what the current
page is.  Of course a few of the pages don't actually have links in the
nav bar, so for them it displays nothing.  Not ideal, but not sure how
to overcome this at the moment.