Moved requests.coffee code into app.coffee
Previously I had an entire file requests.coffee to define the standalone
AJAX requests used in the app.  Currently there's only one, and I don't
forsee too many others, so it seems a bit pointless dedicating an entire
file to 4 or so lines of code.

This changeset just moves that code into an initializer in the
app.coffee file (this is where it was being called from originally
anyway)