Resumr now supports running in systemtest mode
This is a new mode that allows the systemtests to manipulate the server,
and sets up various config options to be suitable for tests

Currently, there are 2 systemtest specific URLs, one which resets the
systemtest git database, and the other which (temporarily) enables
logins.  By default, logins are not required in system test mode.

Also updated the cucumber env.rb file with some corrections, and code to
run the systemtest reset at the end of each test