Fixed issue with spurious section commits
Prior to this changeset, every single UpdateSection call would result in
a commit of the section content, and a commit of the section position.
Since the UI doesn't presently stop the user from hitting save on a
section multiple times this could cause lots of spurious commits to be
created for a section very easily.  Updating the position of a section
also causes a call to this API method, causing even more spurious
commits.

This changeset updates the UpdateSection method to only update the
content/position if they have actually changed.