Split out "Content" functionality from Section
This changeset refactors the Set/Current/ContentHistory functionality
out of the Section class and into it's own base class Content, to allow
the same code to be used for both sections & stylesheets.

Also split out the tests of that functionality into a ContentTests
class, that the tests for Content subclasses can subclass.  ContentTests
contains some metaclass hackery to stop the ContentTests class itself
being picked up as tests to run.