Removed use of a metaclass from base ContentTest
Previously, ContentTest used a meta-class to ensure that subclasses
would be picked up when running tests, but it would not.  Unfortunately,
I only tested this with the standard python unit test runner, and not
with nose.

Nose scans test classes without instantiating them, so the meta-class
was never in effect when building the list of tests to run, so nose
would never run the ContentTest subclass tests.

This changeset fixes that by removing the meta-class and renaming the
test methods to normal test method names.  The contenttests.py file has
been renamed to just content.py so it is not picked up when nose is
scanning for test files.  Unfortunately this will probably cause
failures if the normal unittest runner is used, but at the moment I
think nose is more important.