Refactored creation of content into Content class
Prior to this changeset the document class was responsible for the
initial creation of section data and branches.  This didn't make too
much sense since it meant the Document had to have specific knowledge
of Section internals (how to build ref names etc.).  Sections were also
capable of updating their data, so there was a bit of duplicate/mixed
responsibilities going on.

This changeset shifts all the content creation code into the Content
base class, with a create parameter passed into the constructor.  This
somewhat mirrors the behaviour of the document itself.

Might be worth noting that creation only actually occurs when the
Section is used (or the Create function is called explicitly).  This is
to save on looking up ref/commit data before it is actually needed.