Added CodeMirror & used it for stylesheet editor
This changeset adds the CodeMirror javascript library to provide a code
editor.  The library is currently wrapped with an AMD wrapper and only
provides the CSS editing mode.

A barebones CodeMirror instance has also been added to the stylesheet
editor.  Currently it doesn't do much, or even look particularly good.