Added non-functional preview to stylesheet editor
This commit provides the template for the stylesheet editor preview.
The stylesheet editor is now rendered using a new "editor" layout, with
a bunch of panes that can be used for each of the individual sections of
the editor.  The idea being that the section editor can also be updated
to make use of this layout in the future.

This changeset also updates the sectionList & sectionItem views with
some settings that disable the tools & mouse-over border on each of the
sections, to allow the view to be used in the stylesheet editor.