Refactored css update code into it's own class
Css updating is now handled in a fairly generic way by a new class
called Styler.  The stylesheet editor now fires an event when the
preview needs updated, which will be linked in to the styler's update
function.

This paves the way for using the Styler to update the CSS elsewhere.