Refactored section editor to use editor layout
This commit refactors the section editor to use the editor layout
developed for the stylesheet editor.

It also fixes a whole bunch of problems with the HTML template for the
section editor (think I must have passed the ballmer peak when I wrote
that) which fixes some wierd rendering issues I was getting with some
window widths/layouts.