Fixed bug in pagedown sanitizer
The pagedown sanitizer was removing h4, h5 & h6 tags, even though
markdown actually supports these tags.  This was resulting in one of the
system tests failing, and also rendering problems in certain parts of
the UI (when the pagedown sanitizer was being used).

This commit adds these tags to the whitelist, and fixes the problem