Markdown content is now validated & cleaned
This commit introduces some validation of incoming markdown - any
invalid markdown will be rejected by the server.  Currently this means
any markdown with anything resembling an HTML tag in it.  The rules
might be refined a bit in the future if this proves to be a difficulty.

This changeset also updates the render view to "clean" the markdown
before it is rendered.  This involves stripping out any HTML tags that
are not in the whitelist.