Moved sections api out into seperate blueprint
This commit moves the sections api out into it's own blueprint in the
views/api package.  Also seperated it's tests out into the views/tests
package, and tidied up some small bits of the tests.

Also added flask-should-dsl dependency and started using it for some
bits of the sections test